v0.3 - Categorias, clientes e produtos completos; Iniciando settagem das vendas

+++ Entidades/Produto.cs
@@ -13,7 +13,7 @@ namespace SistemaVenda.Entidades
         [Key]
         public int? Codigo { get; set; }
         public string Descricao { get; set; }
-        public float Quantidade { get; set; }
+        public double Quantidade { get; set; }
         public decimal Valor { get; set; }
 
         [ForeignKey("Categoria")]
